miner finished, can read and insert all metadata

+++ src/models/albums.rs
@@ -1,6 +1,6 @@
 #[derive(Debug)]
 pub struct Albums{
-	id_album: i32,
+	id_album: i64,
 	path: String,
 	name: String,
 	year: u32,
@@ -23,7 +23,7 @@ impl Albums {
 			year: 2000,
 		}
 	}
-	pub fn get_id(&self) -> &i32{
+	pub fn get_id(&self) -> &i64{
 		&self.id_album
 	}
 	pub fn get_path(&self) -> &String{
@@ -38,7 +38,7 @@ impl Albums {
 	pub fn set_path(&mut self, path: String){
 		self.path = path;
 	}
-	pub fn set_id_album(&mut self, id_album: i32){
+	pub fn set_id_album(&mut self, id_album: i64){
 		self.id_album = id_album;
 	}
 	pub fn set_name(&mut self, name: String){
